move dhis2 helper methods to utils
